Ellipse simulation complete
Finished the ellipse simulation code and folded it into the demo.py script.
